+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 - Remove deprecated PublicKey.programId, use SystemProgram.id instead.
 - Remove deprecated PublicKey.ownerValidationProgramId, use OwnerValidationProgram.id instead.
 - Remove deprecated PublicKey.splAssociatedTokenAccountProgramId, use AssociatedTokenProgram.id instead.
-- Remove deprecated typealias AccountInfo, use SPLTokenAccountState or Token2022AccountState instead.
-- Remove deprecated typealias Mint, use SPLTokenMintState or Token2022MintState instead.
+- Remove deprecated typealias AccountInfo, use TokenAccountState or Token2022AccountState instead.
+- Remove deprecated typealias Mint, use TokenMintState or Token2022MintState instead.
 - Remove deprecated typealias Wallet, use AccountBalance instead.
 - Support token 2022 via method getAccountBalances (See GetAccountBalancesTests).
 - Support token 2022 and Token2022Program.